requirements.txt
Django==2.0.1
zappa==0.45.1
djangorestframework==3.7.7
django-model-utils==3.1.1
django-cors-headers==2.1.0
mysqlclient==1.3.7
psycopg2==2.7.3.2
django-storages==1.6.5
pynamodb==3.2.1
django-versatileimagefield==1.8.1
wikipedia==1.4.0
boto3==1.5.22
